+++ apps/server/src/core/error/index.ts
@@ -1,0 +1,1 @@
+export * from './error.module';
