+++ apps/server/src/core/interceptor/index.ts
@@ -1,0 +1,1 @@
+export * from './interceptor.module';
