+++ apps/server/src/core/validation/index.ts
@@ -1,0 +1,1 @@
+export * from './validation.module';
