+++ apps/server/src/modules/authorization/index.ts
@@ -1,0 +1,1 @@
+export * from './authorization.module';
