+++ apps/server/src/modules/files/files.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { FileStorageAdapter } from '@shared/infra/filestorage';
 import { FilesRepo, StorageProviderRepo } from '@shared/repo';
-import { LoggerModule } from '@src/core/logger/logger.module';
+import { LoggerModule } from '@src/core/logger';
 import { DeleteFilesConsole } from './job/delete-files.console';
 import { DeleteFilesUc } from './uc';
