+++ apps/server/src/modules/news/controller/news.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { LoggerModule } from '@src/core/logger/logger.module';
-import { AuthorizationModule } from '@src/modules/authorization/authorization.module';
+import { LoggerModule } from '@src/core/logger';
+import { AuthorizationModule } from '@src/modules/authorization';
 import { NewsRepo } from '@shared/repo';
 import { NewsController } from './news.controller';
 import { NewsUc } from '../uc';
